It works
It's 5:03 am at a hackathon and now it works.
I'm sorry to the future people who are reading this bad commit message.
This is a really useful tool and lots of the design decisions and
reasons for why I have lots of bad practices... good luck :(
Future me is not happy with current me.